refactor settings service and routes
